Link HA data import, geo-ref, switch from 32 to 64 bit (goeref issues)

+++ src/Mapping/SummariseOsData.jl
@@ -28,7 +28,7 @@ function OS_TileSummries(env)
     )
 
 
-    tileRegister_orig = CSV.read(joinpath(env.paths[:projects],".OS_TileRegister.csv"), DataFrame)  
+    tileRegister_orig = CSV.read(joinpath(env.paths[:projects],".OS_TileRegister.csv"), DataFrame)
     newTiles = setdiff(tileRefs[iChecked],tileRegister_orig.tileRefs)
     if !isempty(newTiles)
         iChecked = iChecked[findall(tileRefs[iChecked].==newTiles)]
@@ -36,14 +36,14 @@ function OS_TileSummries(env)
         tileRegister_new = DataFrame(
             :tileRefDirs => join.(dirContents[iChecked],"_"),
             :tileRefs => tileRefs[iChecked],
-            :eastings_max => fill!(Array{Float32}(undef,length(iChecked)),-999.),
-            :northings_max => fill!(Array{Float32}(undef,length(iChecked)),-999.),
-            :eastings_min => fill!(Array{Float32}(undef,length(iChecked)),-999.),
-            :northings_min => fill!(Array{Float32}(undef,length(iChecked)),-999.),
-            :northings_mid => fill!(Array{Float32}(undef,length(iChecked)),-999.),
-            :eastings_mid => fill!(Array{Float32}(undef,length(iChecked)),-999.),
-            :Lat_mid => fill!(Array{Float32}(undef,length(iChecked)),-999.),
-            :Lon_mid => fill!(Array{Float32}(undef,length(iChecked)),-999.)
+            :eastings_max => fill!(Array{Float64}(undef,length(iChecked)),-999.),
+            :northings_max => fill!(Array{Float64}(undef,length(iChecked)),-999.),
+            :eastings_min => fill!(Array{Float64}(undef,length(iChecked)),-999.),
+            :northings_min => fill!(Array{Float64}(undef,length(iChecked)),-999.),
+            :northings_mid => fill!(Array{Float64}(undef,length(iChecked)),-999.),
+            :eastings_mid => fill!(Array{Float64}(undef,length(iChecked)),-999.),
+            :Lat_mid => fill!(Array{Float64}(undef,length(iChecked)),-999.),
+            :Lon_mid => fill!(Array{Float64}(undef,length(iChecked)),-999.)
         )
 
         [tileRegister_new[row,[:eastings_max,:northings_max,:eastings_min,:northings_min,:northings_mid,:eastings_mid]] = MidGridRefs(env,tileRegister_new.tileRefs[row])
